sum constructor

Make a Transformation that computes the sum of bounded data. Use make_clamp to bound data.

Usage

make_sum(input_domain, input_metric)

Arguments

input_domain

Domain of the input data.

input_metric

One of SymmetricDistance or InsertDeleteDistance.

Value

Transformation

Details

If dataset size is known, uses a restricted-sensitivity proof that takes advantage of known dataset size for better utility.

make_sum in Rust documentation.

Citations:

CSVW22 Widespread Underestimation of Sensitivity...

DMNS06 Calibrating Noise to Sensitivity in Private Data Analysis

Supporting Elements:

Input Domain: VectorDomain<AtomDomain<T>>

Output Domain: AtomDomain<T>

Input Metric: MI

Output Metric: AbsoluteDistance<T>